Fixed pre-determained interest rates

+++ Bank.Domain/Models/AccountModel.cs
@@ -11,6 +11,6 @@ namespace Bank.Domain.Models
         public int AccountNumber { get; set; }
         public int TitleID { get; set; }
         public int Balance { get; set; }
-        public Decimal InterestRate { get; set; }
+        public double InterestRate { get; set; }
     }
 }
